Incorporate Build into Elafros.
This copies over the Build type definitions and generates clients for them.

This incorporates BuildSpec into RevisionTemplate and BuildName into Revision.

A RevisionTemplate with a BuildSpec will stamp out a Build during reconciliation, and feed the resulting BuildName to the Revision.

The Revision will reflect the Build's progress in its Status, and stand up resources upon completion.

Fixes: https://github.com/google/elafros/issues/7