Add initial README docs.
These docs are focused on how to start developing for Elafros.

We should probably aim to simplify this to just a couple of steps,
e.g. the k8s docs
(https://github.com/kubernetes/kubernetes#to-start-developing-kubernetes)
seem to make setup look like just a couple of commands (of course
they link to a lot of docs after the initial Go and Docker
configuration).

I've tried to make these docs very beginner friendly, which I think
is a good foot to start off on.

Fixes #28