Stand up Build in "everything", and add Build to several samples.
This adds Steren's two samples, with Elafros objects that build them from source.

This adds a build stanza to Mark's rester-tester sample to build it from source.

This updates the version of the Build type definitions under pkg/apis/cloudbuild, and adds a test that what we have imported matches the version from "@buildcrd//:everything".