Make sure docs explain SSH key setup.
Bazel will be trying to checkout repos it depends on, and it is
currently configured to assume SSH key access, so we should make
sure the docs explain to do this.

Also the list of tasks you have to perform to get ready for
development was missing a bullet point for forking the repo.

Trying to make all of this more clear for folks newer to github also.
I mean GitHub.