Add initial conformance test.
Conformance tests are tests which can be run against any implementation
of the Elafros API to ensure the API has been implemented consistently.
Passing these tests would indicate that apps and functions deployed to
this implementation could be ported to other implementations as well.

The precedent for these tests is the k8s conformance tests which
vendors use to prove they have "certified kubernetes" deployments.
(https://github.com/cncf/k8s-conformance#certified-kubernetes)

This commit includes one conformance test which deploys a new
Route and Configuration from a prebuilt image, then updates the
Configuration to use a different prebuilt image.

Automation to run these tests will be coming later.

The current solution to allowing bazel to run these tests is to
configure bazel to use an empty kubeconfig file. This is pretty
hacky and needs to be revisited later - if we use bazel to run
our automated tests, we might put some meaningful values in this file.

Fixes #89.