Add owners for tests.
@adrcunha @srinivashegde86 @steuhs are driving the test infracstructure
for Elafros, so they should be reviewing and approving changes to the
test directory.

I will be owning the conformance tests and so would like to review
changes to these tests. I've added @adrcunha as well so that if I'm out
there is a backup for these :D

This does not change ownership for the tests in `pkg`, those are
still owned by @evankanderson @mattmoor and @vaikas-google.

Also changed dirs in this file to have leading slashes so they
will be matched only in the root (not dirs with the same name
anywhere in the repo).