Do not update configuration is LatestReadyRevisionName is up to date (#329)
* add islatestuptodate

* change test

* add tests

* delete unused package

* remove comment

* rename test functions

* improve test

* Convert test to call syncHandler directly

Instead of starting the controller and informers, call syncHandler when
necessary in the test. Use Get() on the fake client to test the
state of the Configuration and ensure it hasn't been changed by
addRevisionEvent.

* Clean up test a bit

Expand comments, remove unnecessary field.

* Rewrite test to not use hooks

With informers stopped, call syncHandler directly. Since
addRevisionEvent uses the controller's Lister to get the Configuration
the test needs to add the Configuration to the informer's Indexer
manually.

* Rewrite test to not use hooks

Call addRevisionEvent manually. Since the final state of the
Configuration is not expected to change, we can only test whether
Configuration Update() is called. A reactor fails the test if an update
is detected.

* Correct typo in error message.

* No need to call syncHandler

syncHandler isn't doing anything the test cares about. Only
addRevisionEvent matters.

* Move informer add to Create

Now it's more obvious that this is test setup.

* Update test deps

* fix typo