Rewrite Revision tests to reconcile manually
Due to the lack of fake watches in client-go 1.9, tests that rely on running
informers will fail if all objects are not already present in the client when
informers start. This causes a race and flaky tests, plus tests are
relatively slow since they must wait for informers to start.

Calling syncHandler directly when necessary is more reliable and faster for
testing. In some cases, this is at the expense of test clarity, so we can
revisit these tests when we upgrade to a version of client-go that fixes fake
watches (probably 1.11).

The file queueing_test.go is for testing running controllers, versus
controller_test.go which tests non-running controllers.

Because of the watch issues above, the running controller tests must insert
all objects into the client in advance to avoid racy failures. When these
issues are fixed in client-go these tests can more naturally call Create()
without races.

This change also adopts the shared test helpers and dot import added
in #440.