Rewrite Route tests to call syncHandler directly (#440)
* Rewrite Route tests to call syncHandler directly

Due to the lack of fake watches in client-go 1.9, tests that rely on running
informers will fail if all objects are not already present in the client when
informers start. This causes a race and flaky tests, plus tests are
relatively slow since they must wait for informers to start.

Calling syncHandler and addConfigurationEvent directly when necessary is
more reliable and faster for testing. In some cases, this is at the
expense of test clarity, so we can revisit these tests when we upgrade to
a version of client-go that fixes fake watches (probably 1.11).

The file queueing_test.go is for testing running controllers, versus
controller_test.go which tests non-running controllers.

Because of the watch issues above, the running controller tests must insert
all objects into the client in advance to avoid racy failures. When these
issues are fixed in client-go these tests can more naturally call Create()
without races.

* Generate event delivery expectation functions

Since these functions all have the same structure, they can be generated
by a helper function. This increases the clarity of the test by removing
duplicate lines and naming their intent.

* Correct name of tested object in comment

* Move test helpers to controller/testing package

These helpers are used by all controller tests, so they should only be
defined once. They're now imported as a dot import rather than a named
import to make their use more concise.

* Switch to dot import in route queueing test

For consistency with the rest of the route controller tests.

* Rewrite named routes test to call syncHandler

Necessary due to the lack of fake watches in client-go 1.9.