Use RaceFreeFakeWatcher in ObjectTracker
The FakeWatcher allows sends on the result channel after it's closed,
for example calling Stop() then Add() will panic. RaceFreeFakeWatcher
checks whether the watcher is stopped before attempting to send. It also
panics instead of blocking when the result channel is full.

This commit is copied from kubernetes/kubernetes#61195 where it is
currently unmerged. It's necessary to avoid our tests panicking due
to sends on closed channels.