Do not assert against Generation in conformance tests.
The values the conformance test was asserting against were wrong.
Route.Generation and Configuration.Generation should actually have
a value of 1 when they are first created, and the *.Status.Generation
values are just hacks until CRD generations are updated properly
(which is fixed in https://github.com/kubernetes/kubernetes/issues/58778
but we are not yet using).

I want to remove these assertions since they are just asserting
against a hack. We should actually assert against
Configuration.Generation and Route.Generation when we have pulled in the
fix.