Patch Kubernetes client when updating deps
This patch applied here is the response body of
https://github.com/kubernetes/kubernetes/pull/61195.patch, and also the
same patch as applied in the parent commit. The patch is necessary to avoid
panics when watches fire in Kubernetes fake clients.

Dep will revert the patch every time it runs, so we need to re-patch in
hack/update-deps.sh immediately after dep finishes.