Refactor webhook to allow setting defaults on CRD objects
This updates the GenericCRDHandler to rename the "Callback" function to
"Validator", and adds another function, "Defaultor", to be used for
setting defaults on the objects. The Defaultor is optional.

This also updates the revision webhook to set the
Revision.Spec.ServingState default to RevisionServingStateActive, adds a
test for said behavior, and updates the revision test to use that as
their default value. This is necessary for 0->1/1->0 autoscaling work.