Break concurrency into discrete 100 ms buckets. (#469)
Concurrency is calculated by each individual pod every second and reported to the autoscaler. It reports the number of requests that are in the pod at that instant. And the autoscaler uses those value to determine how many pods the deployment should have.

When requests are very short they will often be missed by the pod's concurrency report, which emits 0. This causes the autoscaler to underestimate the number of pods needed to serve the current traffic load so it scales way down. Then concurrency shoots back up because requests start showing up more often due to longer latency in overtaxed pods. This creates an unstable situation in which pod count flaps and the autoscaler repeatedly panics.

This change puts requests into 100 ms discrete concurrency buckets so that short requests don't go unnoticed. Then the pod reports the average bucket size to the autoscaler. With a target concurrency of 1.0, this allows pods to service 10 qps when latency is very low. Or 1 long-running request at a time. This is a better balance between workload types.