Fix comment on RevisionLabelKey
While looking through recent changes that should be reflected in
the Conformance tests, I found this change. The comment made it
sound like this was a label I should expect to see on a Revision,
so I though the conformance tests needed updating, but after
some digging, it looks like this label is set via
`MakeElaResourceLabels` in ela_resource.go, which is used to set
labels on k8s resources such as (k8s) Services and Deployments.