associate commit id with PR for coverage prow jobs
1. Restructured the way to place the coverage profile for each pull request
build and master commits. So that the profile for each pull build is
under the directory of its PR number. This allows us to find all
commits under a given PR easily
2. Added a head file to record the last pull request build's commit
number. So that we can find the latest build of a pull request easily
3. Added a file that keeps the history of pull request builds  in
chronological order so that the previous commit of any given commit
in a PR can be traced