Add release note block to PR template (#586)
This is inspired by the Kubernetes PR template's release note block. The
intent is that release managers (or a bot) can collate these notes into
the changelog on release. The `release-note` syntax annotation is not a
real syntax, but is intended to make this easier to automate by scanning
PR bodies.