Disallow changes to the Revision Spec.
This change adds logic to validate that the bulk of the Revision's "Spec" field remains unchanged.

If you attempt to change a field of the revision, `kubectl` will now emit an error like:
```shell
...
for: "foo.yaml": Internal error occurred: admission webhook "webhook.elafros.dev" denied the request: mutation failed: Revision spec should not change (-old +new): {v1alpha1.RevisionSpec}.Container.Env[0].Value:
	-: "shiniestnewestversion"
	+: "do not change stuff."
```

The sole field that is currently allowed to change is the `ServingState`, which we expect the autoscaler to use to drive 0->1->0 scaling.

Fixes: https://github.com/elafros/elafros/issues/615