Remove Constructor interface
We'd like to move command flags into the cmd package so controller code isn't
aware of flag parsing, which means constructor signatures need to be updated to
allow flag values to be passed in. This requires removing the Constructor
interface to allow constructor signatures to vary depending on which flags are
relevant to each controller.

The original idea was that a constructor interface was useful because we'd be
adding lots of controllers, but it turned out adding a controller is a rare
occurrence.