Fix the race condition in Configuration updates Fixes #710
This fixes the irreconcilable state issue due to a race condition
in configuration updates from route controller and configuration
controller. If route controller succeeds in updating the configuration
first, it gets into a state where revision is created but configuration
status is not updated. This causes the synchandler to return error each
time when it is called.