Remove github.com/evanphx/json-patch dependency (#731)
The only reason for adding this dependency was so we could use the
merge patch functionality in the conformance test, but there is no
reason the test couldn't instead use JSONPatch, which is a feature
of `mattbaird/jsonpatch` which we actually need for core Elafros
functionality.

I also looked into switching from `mattbaird/jsonpatch` to
`evanphx/json-patch`, which was appealing because `evanphx/json-patch`
has good signals such as > stars, godocs, CI, A+ go report, however
it turns out the interface in `mattbaird/jsonpatch` for declaring
patches is nicer (a typed interface v.s. writing raw json).