Vendor elafros/build.
Previously, we had a copy of the elafros/build types inlined under pkg/apis/build, from which we generated client libraries ourselves.

This change vendors the elafros/build repository and instead relies on the client libraries generated and checked into that repository.

This makes progress on #270, but doesn't complete it because we still rely on `@buildcrd//:everything` through Bazel for the contributor workflow.  I will address this in a subsequent change.