Complete the vendoring of elafros/build.
This completes the process from #829 and removes the WORKSPACE dependency on elafros/build. We now rely on a copy of the release.yaml that was published from elafros/build's nightly release.

This also updates the revision of Build that we depend on to reflect the release.yaml we are pulling in.

Fixes: #270