Add tag-to-digest resolution as part of Elafros deployment. (#827)
As part of turning the Revision => Deployment, resolve tags to digests. This uses the github.com/google/go-containerregistry client to access the image on the remote repository and the extension github.com/mattmoor/k8schain, which provided a Kubelet compatible auth keychain for use with the library. This makes the tag-to-digest resolution code itself remarkably small, and me really happy.

With this change, we're able to start surfacing [error conditions](https://github.com/elafros/elafros/blob/master/docs/spec/errors.md#container-image-not-present-in-repository) accessing the image.  e.g. if I change the `python-simple` sample to use the tag `:fakest` GCR returns the error:
```yaml
  status:
    conditions:
    - lastTransitionTime: 2018-05-08T03:29:54Z
      message: 'MANIFEST_UNKNOWN: "Failed to fetch \"fakest\" from request \"/v2/convoy-adapter/python-simple/manifests/fakest\"."'
      reason: ContainerMissing
      status: "False"
      type: ContainerHealthy
    - lastTransitionTime: 2018-05-08T03:29:54Z
      message: 'MANIFEST_UNKNOWN: "Failed to fetch \"fakest\" from request \"/v2/convoy-adapter/python-simple/manifests/fakest\"."'
      reason: ContainerMissing
      status: "False"
      type: Ready
```

Fixes: https://github.com/elafros/elafros/issues/498