Remove `ServingState` from `RevisionTemplateSpec` requests (#893)
The `RevisionTemplateSpec`, which the user specifies when creating a
`Configuration`, contains the `ServingState` field because it contains
all of the fields that the `RevisionSpec` contains. `RevisionSpec` is
created internally from the `RevisionTemplateSpec` and the user should
not be providing a value for `ServingState` as this will be provided
internally when the `RevisionSpec` is created.

However since the field did not have `omitempty`, it was actually being
provided in requests from Elafros clients with `""` as the value. Since
we never want the user to provide a value for `ServingState`, this adds
`omitempty` so that the field will not be included in the request.