Use json field `type` for Type instead of `state` (#902)
How did this happen you wonder?

I'm glad you asked!

Before the dawn of time as far as git is concerned) this was
implemented as a custome api server. Back then, this bug existed:
https://github.com/kubernetes-incubator/apiserver-builder/issues/176

This meant that instead of defining custom types such as we are using
(`RouteConditionType`), one had to use `string` explicitly. At that
time, a good soul added a comment saying when this was fixed, we should
use a line which specifyed a custom type.

BUT LO AND BEHOLD THAT COMMENT CONTAINED A TYPO! Instead of `type`,
it used `state` as the json field!

When the move was made from custom api server to CRDs, the line was
dutifully obeyed, uncommented and BOOM we were serializing and
deserializing the `Type` field as `State`.

This was then copied into the `Service` implementation.

As for why this was ever `state` in the comment to start with, that
mystery may never be solved.

The end!