Add a patchBase parameter to ResourceDefaulter
This enables patches to be rooted at something other than "/spec/. This
is needed to implement the Service ResourceDefaulter, which will
delegate to the Configuration ResourceDefaulter to set its defaults.

Concretely, when setting the default for Configuration, the patchBase
passed to SetConfigurationDefaults will be "/spec/"; when setting the
default for a Configuration embedded in a Service, the patchBase passed
to SetConfigurationdefaults will be "/spec/pinned/configuration/" or
"/spec/runLatest/configuration/".