Rename the CRD group to serving.knative.dev
Other Knative projects use the project name in the CRD group. Now that
the Serving project is no longer called "elafros", the project name
should be included in the CRD group namespace.

This is a breaking change, but aligns well with the other breakages
caused by the rename.

I intentionally did not update name for resource labels or for logging.

In docs there were a few references to CRDs from the build project.
Those CRD groups are now called build.knative.dev.