Use unique names in end to end tests
The end to end test (both `e2e` and `conformance`) were using hardcoded
names for the `Route` and `Configuration` objects they create. This
means trouble if anyone were to run the same tests against the same
cluster. It also means if a test fails to clean up after itself, it
interferes with the next test.

Now the names will be randomized; one downside is that it might skip
issues that could happen trying to recreate resources with the same name
but if we want to test that I think we should add tests specifically for
that case.

* Also updated the `CleanupOnInterrupt` method to actually exist,
  otherwise it actually swallows the interrupt and the main goroutine just
  keeps on truckin.
* Also renamed `e2e_test.go` to `e2e.go` since it doesn't contain any tests
* Also fixed some formatting in adding_tests.md

_An alternative to randomly generating these names in the test would be
[to let the server do
it](https://github.com/kubernetes/community/blob/master/contributors/devel/api-conventions.md#idempotency)
however it seems more like what a client could do to specify the name
directly so I'm going with that for now._