Refactor autoscaler for future multitenancy
These changes make the autoscaler code ready for a future refactor to a
multitenant autoscaling controller.

The autoscaler Config now takes standard Go variables instead of
k8sflag.

The struct sent over the websocket is now a wrapper that includes the
revision key (namespace/name) along with the Stat. The ELA_REVISION env
var is now the revision key instead of the revision name to uniquely
identify a revision.

Autoscaler instances are now thread-safe.

StatReporter metrics setup is now done in init() instead of on creation,
allowing multiple StatReporters to be created in the same process.

The autoscalerImage flag is no longer required. If not present, no
single-tenant autoscaler deployment and service will be created.

The queue proxy will now read from the websocket connection to receive
Close messages. This allows the queue proxy to reconnect much more
quickly (5 seconds vs 1 minute in my testing).