Only run e2e tests when -tag=e2e is specified (#1053)
As part of the move to ko (and generally improving the dev experience)
@mattmoor would like `go test ./...` to run only the unit tests and
not the end to end tests. One neat way to accomplish that is to use
[build constraints](http://peter.bourgon.org/go-in-production/#testing-and-validation).
(One downside here is that folks adding new e2e tests need to
remember to add the constraint.)

I also tried to de-dupe the docs on how to run tests since they're
pretty much the same for conformance tests and e2e tests. It's a bit
confusing that conformance tests are also e2e tests but maybe it's
okay? ¯\_(ツ)_/¯

Fixes #1028