Move revision condition manipulation to _types.go (#1118)
* Move revision condition manipulation into revision_types.go

This is the Revision equivalent of #1113

This contains two other unrelated changes:
1. Condense the object construction in `revision_test.go`
1. Reorder the `helloworld` yaml to create the configuration first.

The reason for the second "unrelated" change is that I was seeing the route controller marking the Route as "AllTrafficAssigned = False" because the configuration wasn't created yet.

* Update the conditions spec to eliminate wording about elision.

Eliding conditions can come later, it's a premature optimization,
and I'm not sure we're handling conditions well enough to do it
reliably yet.

* Fix test breakages due to setCondition becoming private.

* Add unit test coverage of the condition logic.

* Clarify isBuildDone through local variable names.