Fix ProgressDeadlineExceeded handling. (#1166)
The manner in which we were looking up the controlling Revision on Deployment events was relying on (an incorrect) naming convention where `deployment.Name == revision.Name`.  In reality, `deployment.Name == revision.Name + "-deployment"`, but relying on this for looking up the controlling object is an antipattern.

Instead, we should utilize `metav1.GetControllerOf` to determine the controlling `OwnerReference` and use this name to lookup the owning `Revision.

This fixes the local issue, but I've opened #1165 to track a broader triage of this.

Progress toward: https://github.com/knative/serving/issues/1150