Create Check methods parallel to WaitFor methods.
Every check we perform shouldn't be a WaitFor; we should WaitFor readiness and then assert things work in-the-now.  This creates parallel methods for checking things, which has the same signature as their `WaitFor` siblings.

This uses these methods for the recently sunk checks in the routing conformance test.

Related: https://github.com/knative/serving/issues/1178