Improve concurrent request bucketing in queue-proxy. (#1091)
* Improve concurrent request bucketing in queue-proxy.

Instead of counting all requests that arrived in a certain bucket as concurrent, the queue proxy now reports the actual maximum concurrency that was present inside of one bucket.

If for example 3 requests arrive at once, the maximum concurrency will be 3 for that bucket. If another arises while the 3 remaining are still open, maximum concurrency is 4. Closing requests results in a decrement on the concurrency immediatly (versus draining the outgoing request queue on quantization which results in the behavior described above).

Closes #1060

* Fix edge case and add a specific test.

* Unify channels and add buffer again.

* Update config documentation.

* Adjust channel typing.