Add logic to avoid overwriting conditions in the correct state. (#1184)
Prior to the change that disabled `LastTransitionTime` calls to `setCondition` would first remove the prior condition of a given type.  Then it would stamp the new condition with the current timestamp.  Finally it would append the condition to the list.  With this, it was possible that the only thing changing in the condition was the timestamp, leading to constant reconciliation.

This change adds a check to the loop that removes any old conditions.  If the new condition with the old condition's timestamp is identical, then the `setCondition` is simply elided as a nop.  This causes reconciliation the close relatively quickly.

Fixes: https://github.com/knative/serving/issues/1181