Restructure the base/controller split.
This is the first change of several that will restructure how our controllers process work.  This change simply reworks the Base constructor interface to make it harder (I think) to not add an informer to the set we are syncing before we start reconciliation.

This also moves the burden of adding the primary resource's event handler back into the individual controllers (as a prelude to making our other events go through the same workqueue).