Refactor the Configuration controller to use level-based reconciliation.
Based on the guidance from Clayton and others at today's working group meeting, this starts to move our controllers towards more conventional reconciliation.  Previously, we did work in the callbacks from informers directly.  With this, we still listen to these secondary events, but when they occur instead of taking immediate action, we instead queue a reconciliation of any primary resources with which they are associated.

In simpler terms, instead of having the Configuration controller take direct action when Revision events occur we instead queue a reconciliation of the Controller that created it.  This core reconciliation loop is now responsible for 100% of the logic so there are changes there as well to accommodate this shift.