Move the Service controller to a level-based model.
This moves the service controller to the more level-based model started in #1202.  This also makes the service controller start to listen to Route and Configuration events to trigger reconciliation, so that we can reflect a basic overall readiness condition based on the combination of these two sub-resources readiness conditions (I don't expect this will be the final model, but is a start).

Fixes: https://github.com/knative/serving/issues/1134