Use K8s blessed equality comparison.
This changes the comparison used in the Service controller's `reconcileFoo` methods to use `equality.Semantic.DeepEqual`.

Although it has slightly different semantics (stricter) I left the use of `cmp.Diff` in what we log as the difference when reconciliation is to be performed, since I think this will really help give visibility into the changes we see triggering reconciliation.