Review feedback (#1225)
* Normalize the use of []ConditionType in pkg/apis/serving.

This wraps the slice literals so a single element per-line:
```
[]FooConditionType{
    FooType,
    BarType,
    BazType,
}
```

It also inlines the literal into the `for x, y := range ...` unless it is dynamically constructed (`if foo { append }`) and needs the variable.

* Rename PropagateFoo to PropagateFooStatus.

* Wrap the informer list one-per-line consistently.

* Rename the `informer`s in each controller for consistency.

* Rename the lister fields on controllers to fooLister as appropriate.

* Collapse the client construction into calls through it.

* Use K8s blessed equality comparison.

This changes the comparison used in the Service controller's `reconcileFoo` methods to use `equality.Semantic.DeepEqual`.

Although it has slightly different semantics (stricter) I left the use of `cmp.Diff` in what we log as the difference when reconciliation is to be performed, since I think this will really help give visibility into the changes we see triggering reconciliation.