Add Revision defaulter
Defaulters are a feature of runtime.Scheme that can set default values
for any runtime.Object. We can use kubernetes/code-generator's
defaulter-gen to generate defaulters for use with CRDs like Revision.

The k8s:defaulter-gen=TypeMeta comment tells defaulter-gen to generate
defaulters for every struct with a TypeMeta field. In this case, we want
to set the ServingState of a RevisionSpec to Active by default, so we
need a defaulter for Revision since RevisionSpec doesn't embed TypeMeta.

The actual defaulting function is SetDefault_Revision. The generated code
makes it easy to hook this function into the Scheme.

To set defaults on an object, import the clientset's generated scheme
package and call scheme.Scheme.Default(object). The test can't do this
because of an import cycle, so it creates a new scheme.