Create informers in cmd/controller/main.go (#1230)
This change hoists the creation of our informers into our controller's main.go and passes into each controller the set of informers it uses.  This enables us to move the logic to wait for the informers to sync into main.go as well, and (hopefully) make it harder for folks to get wrong.

This also moves `BuildClientSet` into `controller.Options` alongside its siblings and renames `ElaClientSet` to `ServingClientSet.