exclude test utility code from coverage testgrid
Coverage testgrid uses the github attribute linguist-generated=true to exclude items. This commit is to fix the issue https://github.com/knative/serving/issues/1122.

There may be side effects, because I am not sure whether other parties rely on the "linguist-generated" attribute as well and the change here may not be desired for them.