exclude test utility code from coverage testgrid (#1256)
* exclude test utility code from coverage testgrid

Coverage testgrid uses the github attribute linguist-generated=true to exclude items. This commit is to fix the issue https://github.com/knative/serving/issues/1122.

There may be side effects, because I am not sure whether other parties rely on the "linguist-generated" attribute as well and the change here may not be desired for them.

* exclude test utility code from coverage testgrid

This commit is a step toward fixing the issue https://github.com/knative/serving/issues/1122. We add the
coverage-excluded to items we want to exclude from code coverage calculation & display

* exclude test utility code from coverage testgrid

This commit is a step toward fixing the issue https://github.com/knative/serving/issues/1122. We add the
coverage-excluded to items we want to exclude from code coverage calculation & display

* makes sure pkg/controller/testing folder is still reviewable