Start migrating the Revision controller to level-based Reconciliation. (#1268)
This moves the very first part of the Revision controller's reconciliation to be level-based.  When a Revision has an associated Build, the Revision fetches the state of that Build and incorporates its status into its own.  Then, depending on the result of the Build, it proceeds with the reconciliation of other resources.  In a pivot to being more level-based, when Build events trigger we `Enqueue` the Revisions to reconcile instead of directly processing the events.