Give our resource deletion calls a more regular shape. (#1305)
A subset of these calls were calling Get before Delete, vs. just using "not found" on the deletion to elide logging.  I hoisted the foreground `DeleteOptions` into a shared variable to cut down the boilerplate some.

This also removes a refresh I speculatively added in a prior change in an attempt to combat optimistic concurrency failures (not convinced it's really doing anything but adding complexity).