Always call updateStatus after Reconcile. (#1321)
* Always call updateStatus after Reconcile.

This splits the main `Reconcile` methods of the controllers, so that regardless of whether they return an error we can call `updateStatus` with any changes made to the object being reconciled regardless of the success or failure of the reconciliation.  This also removes all of the `updateStatus` calls from the various methods with the exception of the methods called directly by the informer.

Related: https://github.com/knative/serving/issues/1107

* Add back logging to the updateStatus failures that lost it.