This hoists the K8s Service reconciliation and makes it level-based.
This is the next phase of our level-based migration of the Revision controller.  It includes the K8s Service for the user Deployment. This is patterned in much the same way as the Deployment refactor in #1322.  The most notable difference in this change is that we establish readiness from the Endpoints resource instead of the Service resource directly.

In addition, I spent a bit of time cleaning up some of the `revision_test.go` code to more automatically update the informers as resources are reconciled.  In particular, after `controller.Reconcile` we need to make sure all of the updates to the underlying objects are reflected in our informers, or a subsequent reconcile may act on stale state.  To facilitate this, I added a helper that updates the informers with the Revision/Deployment/Service to be used after a `controller.Reconcile`.