Move the Configuration controller testing to use the new table-testing method.
This uses the same strategy as the Service controller, which is based on the OpenShift Ingress controller's tests.  Backport a few fixes to the way we set up ObjectMeta and OwnerReferences in our Service tests uncovered bringing this up.  This also changes the way we create Build names to more easily test it (mirrors what we do for Revision today).

The most notable difference from the Service controller testing is that this needs to watch for Actions across 2 clients, so I refactored the loop body a bit to support that.  This should make it easier to copy to Route and Revision.

Fixes: https://github.com/knative/serving/issues/1220