Always output detailed e2e test results
We want to use our end to end tests to take some latency measurements
over time. In our first pass, we want to do this at the granularity
of a test, by tracking the time taken by each test.

To do this, we want the output from running our e2e tests with prow
to include a line with the time taken by a test and the name of a test.

It turns out that running `go test` with `-v` will automatically
include this information, so we basically get it for free, however
we are currently supressing this when the tests succeed and only
displaying it when they fail.

This commit changes it to always output the duration of each
individulal test, even when successful.

One side effect is that the info logged by the tests (i.e. each
"step" of the test) will also be included in successful test runs.
We may want to adjust this in the future as more tests are added
so that we can have verbose output without detailed steps being logged.

Fixes #1008