Make Deployment reconciliation level-based. (#1322)
This starts to push the current `switch` over `ServingState` in the main `Reconcile` method into the methods that reconcile individual resources.  As `reconcileFoo` handles the complete reconciliation of the `Revision` `Foo` resource, we will hoist it above the switch until the `{create,delete}K8sResources` methods are empty, at which point we will delete the switch.