Add support for emitting metrics from tests.
We want to track latency over time by emitting this information from
our end to end tests.

In this change we use opencensus metric collection to collect and
export metrics to the zap logger. The `Wait` (aka polling) functions
used by the end to end tests will automatically emit these metrics,
identified by a string provided by the caller.

These metrics will be emitted if an end to end test is run with
`--emitmetrics`.

Some choices made in this implementation:
* The metric tracker is global, effectively making the logger global,
  which is contrary to the current logger interface (it must be
  passed around)
* The fact that the caller can provide any string is a bit weird,
  but also the metric is not meaningful without some context.