Refactor the way our Controllers watch ConfigMaps
Previously the Revision and Route controllers used a ConfigMap informer directly to watch for changes.  This generalizes that pattern into a slightly higher level abstraction under `pkg/configmap`.

With this `configmap.Watcher` abstraction, Controllers simply register to `Watch` particular `ConfigMap`s for changes via a callback:
```go
  cm.Watch("configmap-name", c.receiveSomeConfig)
```

Once `Start()` is called on the `configmap.Watcher` all of the registered `configmap.Observers` will be invoked with the initial state of their `ConfigMap`s, and if any are not found the call to `Start()` will return an error.  This consolidates the setup and update logic for the various configurations behind a single interface.

This transitions the `NetworkConfig` and `DomainConfig` logic, which already employ this pattern (sans abstraction) to this abstraction.

Fixes: https://github.com/knative/serving/issues/1242