Do not log in log library
Okay this is slightly more repeated code, but hear me out! When you log in
a library, it means the caller has no control over whether those
logs are emitted or not - case in point, the e2e tests are using this
library, and now they dump the json config and log level at the
beginning of every test, which we don't want and can't control.

This commit changes the `logging` lib such that it's the caller's
responsibility to decide whether to log or not.

This also changes the logic such that the logging configuration is no
longer logged at debug.